requirements.txt
# Основные зависимости
gspread==5.11.3
google-auth==2.25.2
requests==2.31.0
schedule==1.2.1

# Для логирования (если нужно расширенное)
# python-json-logger==2.0.7

# Для работы с датами
python-dateutil==2.8.2